Adding JWT claims to deleting items

+++ jwt_extended/resources/item.py
@@ -1,6 +1,6 @@
 """Set up of resources for app."""
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims
 from models.item import ItemModel
 
 
@@ -49,13 +49,18 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
         """DELETE /item route."""
+        claims = get_jwt_claims()
+        if not claims["is_admin"]:
+            return {"message": "Admin privilege required"}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
-
-        return {"message": "Item deleted"}
+            return {"message": "Item deleted"}
+        return {"message": "Item not found"}, 404
 
     def put(self, name):
         """PUT /item route."""
